chore(test): refactor testcase filename
- refactor: testcase filename
  - __tests__/* > **/*.test.ts
