+++ src/ajv/getValidationErrorSummary.test.ts
@@ -1,5 +1,5 @@
 import { getValidationErrorSummary } from '#/ajv/getValidationErrorSummary';
-import { ErrorObject } from 'ajv';
+import type { ErrorObject } from 'ajv';
 import { beforeAll, describe, expect, it } from 'vitest';
 
 const context: {
