+++ src/modules/noop.test.ts
@@ -1,0 +1,8 @@
+import { noop } from '#/modules/noop';
+import { describe, it } from 'vitest';
+
+describe('noop', () => {
+  it('pass', () => {
+    noop();
+  });
+});
