+++ src/stringify/safeStringify.test.ts
@@ -1,19 +1,5 @@
-import { escape } from '#/stringify/escape';
-import { objectify } from '#/stringify/objectify';
 import { safeStringify } from '#/stringify/safeStringify';
 import { describe, expect, it, vitest } from 'vitest';
-
-describe('escape', () => {
-  it('add backslash', () => {
-    const r01 = escape('test\ntest');
-    expect(r01).toEqual('test\\\ntest');
-  });
-
-  it('replace another charactor', () => {
-    const r01 = escape('test\ntest', '1');
-    expect(r01).toEqual('test1test');
-  });
-});
 
 describe('safeStringify', () => {
   it('plain object', () => {
@@ -48,22 +34,5 @@ describe('safeStringify', () => {
   it('replace, space', () => {
     const r01 = safeStringify({ name: 'ironman' }, (_k, v) => v, 2);
     expect(r01).toEqual(`{\n  "name": "ironman"\n}`);
-  });
-});
-
-describe('objectify', () => {
-  it('function termination', () => {
-    const r01 = objectify({ name: 'ironman', f: (name: string) => `${name}:hello?` });
-    expect(r01).toMatchObject({ name: 'ironman' });
-  });
-
-  it('exception', () => {
-    const err = new Error('error');
-    const spyH = vitest.spyOn(JSON, 'stringify').mockImplementationOnce(() => {
-      throw err;
-    });
-    const r01 = objectify({ name: 'ironman', f: (name: string) => `${name}:hello?` });
-    spyH.mockRestore();
-    expect(r01).toEqual({ err });
   });
 });
